==> analysis/08-marker-detection.Rmd <==
---
title: "Marker detection"
author: "James Ashmore"
date: "`r format(Sys.time(), '%d %B, %Y')`"
output: workflowr::wflow_html
editor_options:
  chunk_output_type: console
---

## Motivation

In this analysis we will identify differentially expressed genes between clusters in order to create a set of marker genes.

## Setup

Set chunk options:

```{r}
knitr::opts_chunk$set(
  autodep    = TRUE,
  cache      = TRUE,
  cache.path = "cache/08-marker-detection.Rmd/",
  dev        = "png",
  echo       = TRUE,
  error      = FALSE,
  message    = FALSE,
  warning    = FALSE
)
```

Load required packages:

```{r}
library(scater)
library(scran)
```

Import datatable options:

```{r, include = FALSE}
DT::datatable(matrix())
```

Read experiment data:

```{r}
sce <- readRDS("data/07-cell-cycle.rds")
```

## Welch *t*-test

### Non-specific

Find non-specific candidate marker genes between clusters:

```{r}
res <- findMarkers(sce, sce$cluster, test.type = "t", pval.type = "any", direction = "up")

sig <- lapply(res, subset, FDR < 0.05)

sig <- Filter(nrow, sig)
```

#### Table {.tabset}

Create a table of all significant marker genes for each cluster:

```{r, results = "asis", echo = FALSE}
for (n in names(sig)) {

  dat <- as.data.frame(sig[[n]])

  cat("##### Cluster", n, "\n")

  cat(knitr::knit_print(DT::datatable(dat)))

  cat("\n\n")

}
```

#### Heatmap {.tabset}

Plot heatmap of gene expression values for the Top 10 marker genes:

```{r, fig.height = 8, fig.width = 10, results = "asis"}
for (n in names(sig)) {
  
  dat <- sig[[n]]
  
  top <- head(dat, n = 10)
  
  ids <- rownames(top)

  cat("##### Cluster", n, "\n")
  
  plotHeatmap(sce, colour_columns_by = "cluster", features = ids, center = TRUE, symmetric = TRUE, cluster_rows = ifelse(length(ids) >= 2, TRUE, FALSE), show_colnames = FALSE)
  
  cat("\n\n")
  
}
```

#### Dimension {.tabset}

Plot reduced dimension coloured by gene expression values for the Top 10 marker genes:

```{r, fig.height = 8, fig.width = 10, results = "asis"}
for (n in names(sig)) {

  dat <- sig[[n]]

  top <- head(dat, n = 10)

  ids <- rownames(top)

  plt <- lapply(ids, function(x) plotUMAP(sce, colour_by = x) + ggtitle(x))

  cat("##### Cluster", n, "\n")

  print(patchwork::wrap_plots(plt))

  cat("\n\n")

}
```

#### Expression {.tabset}

Plot distribution of gene expression values for the Top 10 marker genes:

```{r, fig.height = 8, fig.width = 10, results = "asis"}
for (n in names(sig)) {

  dat <- sig[[n]]

  top <- head(dat, n = 10)

  ids <- rownames(top)

  plt <- plotExpression(sce, features = ids, x = "cluster", colour_by = "cluster", scales = "free")

  cat("##### Cluster", n, "\n")

  print(plt)

  cat("\n\n")

}
```

### Specific

Find specific candidate marker genes between clusters:

```{r}
res <- findMarkers(sce, sce$cluster, test.type = "t", pval.type = "all", direction = "up")

sig <- lapply(res, subset, FDR < 0.05)

sig <- Filter(nrow, sig)
```

#### Table {.tabset}

Create a table of all significant marker genes for each cluster:

```{r, results = "asis", echo = FALSE}
for (n in names(sig)) {

  dat <- as.data.frame(sig[[n]])

  cat("##### Cluster", n, "\n")

  cat(knitr::knit_print(DT::datatable(dat)))

  cat("\n\n")

}
```

#### Heatmap {.tabset}

Plot heatmap of gene expression values for the Top 10 marker genes:

```{r, fig.height = 8, fig.width = 10, results = "asis"}
for (n in names(sig)) {
  
  dat <- sig[[n]]
  
  top <- head(dat, n = 10)
  
  ids <- rownames(top)

  cat("##### Cluster", n, "\n")
  
  plotHeatmap(sce, colour_columns_by = "cluster", features = ids, center = TRUE, symmetric = TRUE, cluster_rows = ifelse(length(ids) >= 2, TRUE, FALSE), show_colnames = FALSE)
  
  cat("\n\n")
  
}
```

#### Dimension {.tabset}

Plot reduced dimension coloured by gene expression values for the Top 10 marker genes:

```{r, fig.height = 8, fig.width = 10, results = "asis"}
for (n in names(sig)) {

  dat <- sig[[n]]

  top <- head(dat, n = 10)

  ids <- rownames(top)

  plt <- lapply(ids, function(x) plotUMAP(sce, colour_by = x) + ggtitle(x))

  cat("##### Cluster", n, "\n")

  print(patchwork::wrap_plots(plt))

  cat("\n\n")

}
```

#### Expression {.tabset}

Plot distribution of gene expression values for the Top 10 marker genes:

```{r, fig.height = 8, fig.width = 10, results = "asis"}
for (n in names(sig)) {

  dat <- sig[[n]]

  top <- head(dat, n = 10)

  ids <- rownames(top)

  plt <- plotExpression(sce, features = ids, x = "cluster", colour_by = "cluster", scales = "free")

  cat("##### Cluster", n, "\n")

  print(plt)

  cat("\n\n")

}
```

## Wilcoxon rank sum test

### Non-specific

Find non-specific candidate marker genes between clusters:

```{r}
res <- findMarkers(sce, sce$cluster, test.type = "wilcox", pval.type = "any", direction = "up")

sig <- lapply(res, subset, FDR < 0.05)

sig <- Filter(nrow, sig)
```

#### Table {.tabset}

Create a table of all significant marker genes for each cluster:

```{r, results = "asis", echo = FALSE}
for (n in names(sig)) {

  dat <- as.data.frame(sig[[n]])

  cat("##### Cluster", n, "\n")

  cat(knitr::knit_print(DT::datatable(dat)))

  cat("\n\n")

}
```

#### Heatmap {.tabset}

Plot heatmap of gene expression values for the Top 10 marker genes:

```{r, fig.height = 8, fig.width = 10, results = "asis"}
for (n in names(sig)) {
  
  dat <- sig[[n]]
  
  top <- head(dat, n = 10)
  
  ids <- rownames(top)

  cat("##### Cluster", n, "\n")
  
  plotHeatmap(sce, colour_columns_by = "cluster", features = ids, center = TRUE, symmetric = TRUE, cluster_rows = ifelse(length(ids) >= 2, TRUE, FALSE), show_colnames = FALSE)
  
  cat("\n\n")
  
}
```

#### Dimension {.tabset}

Plot reduced dimension coloured by gene expression values for the Top 10 marker genes:

```{r, fig.height = 8, fig.width = 10, results = "asis"}
for (n in names(sig)) {

  dat <- sig[[n]]

  top <- head(dat, n = 10)

  ids <- rownames(top)

  plt <- lapply(ids, function(x) plotUMAP(sce, colour_by = x) + ggtitle(x))

  cat("##### Cluster", n, "\n")

  print(patchwork::wrap_plots(plt))

  cat("\n\n")

}
```

#### Expression {.tabset}

Plot distribution of gene expression values for the Top 10 marker genes:

```{r, fig.height = 8, fig.width = 10, results = "asis"}
for (n in names(sig)) {

  dat <- sig[[n]]

  top <- head(dat, n = 10)

  ids <- rownames(top)

  plt <- plotExpression(sce, features = ids, x = "cluster", colour_by = "cluster", scales = "free")

  cat("##### Cluster", n, "\n")

  print(plt)

  cat("\n\n")

}
```

### Specific

Find specific candidate marker genes between clusters:

```{r}
res <- findMarkers(sce, sce$cluster, test.type = "wilcox", pval.type = "all", direction = "up")

sig <- lapply(res, subset, FDR < 0.05)

sig <- Filter(nrow, sig)
```

#### Table {.tabset}

Create a table of all significant marker genes for each cluster:

```{r, results = "asis", echo = FALSE}
for (n in names(sig)) {

  dat <- as.data.frame(sig[[n]])

  cat("##### Cluster", n, "\n")

  cat(knitr::knit_print(DT::datatable(dat)))

  cat("\n\n")

}
```

#### Heatmap {.tabset}

Plot heatmap of gene expression values for the Top 10 marker genes:

```{r, fig.height = 8, fig.width = 10, results = "asis"}
for (n in names(sig)) {
  
  dat <- sig[[n]]
  
  top <- head(dat, n = 10)
  
  ids <- rownames(top)

  cat("##### Cluster", n, "\n")
  
  plotHeatmap(sce, colour_columns_by = "cluster", features = ids, center = TRUE, symmetric = TRUE, cluster_rows = ifelse(length(ids) >= 2, TRUE, FALSE), show_colnames = FALSE)
  
  cat("\n\n")
  
}
```

#### Dimension {.tabset}

Plot reduced dimension coloured by gene expression values for the Top 10 marker genes:

```{r, fig.height = 8, fig.width = 10, results = "asis"}
for (n in names(sig)) {

  dat <- sig[[n]]

  top <- head(dat, n = 10)

  ids <- rownames(top)

  plt <- lapply(ids, function(x) plotUMAP(sce, colour_by = x) + ggtitle(x))

  cat("##### Cluster", n, "\n")

  print(patchwork::wrap_plots(plt))

  cat("\n\n")

}
```

#### Expression {.tabset}

Plot distribution of gene expression values for the Top 10 marker genes:

```{r, fig.height = 8, fig.width = 10, results = "asis"}
for (n in names(sig)) {

  dat <- sig[[n]]

  top <- head(dat, n = 10)

  ids <- rownames(top)

  plt <- plotExpression(sce, features = ids, x = "cluster", colour_by = "cluster", scales = "free")

  cat("##### Cluster", n, "\n")

  print(plt)

  cat("\n\n")

}
```

## Binomial test

### Non-specific

Find non-specific candidate marker genes between clusters:

```{r}
res <- findMarkers(sce, sce$cluster, test.type = "binom", pval.type = "any", direction = "up")

sig <- lapply(res, subset, FDR < 0.05)

sig <- Filter(nrow, sig)
```

#### Table {.tabset}

Create a table of all significant marker genes for each cluster:

```{r, results = "asis", echo = FALSE}
for (n in names(sig)) {

  dat <- as.data.frame(sig[[n]])

  cat("##### Cluster", n, "\n")

  cat(knitr::knit_print(DT::datatable(dat)))

  cat("\n\n")

}
```

#### Heatmap {.tabset}

Plot heatmap of gene expression values for the Top 10 marker genes:

```{r, fig.height = 8, fig.width = 10, results = "asis"}
for (n in names(sig)) {
  
  dat <- sig[[n]]
  
  top <- head(dat, n = 10)
  
  ids <- rownames(top)

  cat("##### Cluster", n, "\n")
  
  plotHeatmap(sce, colour_columns_by = "cluster", features = ids, center = TRUE, symmetric = TRUE, cluster_rows = ifelse(length(ids) >= 2, TRUE, FALSE), show_colnames = FALSE)
  
  cat("\n\n")
  
}
```

#### Dimension {.tabset}

Plot reduced dimension coloured by gene expression values for the Top 10 marker genes:

```{r, fig.height = 8, fig.width = 10, results = "asis"}
for (n in names(sig)) {

  dat <- sig[[n]]

  top <- head(dat, n = 10)

  ids <- rownames(top)

  plt <- lapply(ids, function(x) plotUMAP(sce, colour_by = x) + ggtitle(x))

  cat("##### Cluster", n, "\n")

  print(patchwork::wrap_plots(plt))

  cat("\n\n")

}
```

#### Expression {.tabset}

Plot distribution of gene expression values for the Top 10 marker genes:

```{r, fig.height = 8, fig.width = 10, results = "asis"}
for (n in names(sig)) {

  dat <- sig[[n]]

  top <- head(dat, n = 10)

  ids <- rownames(top)

  plt <- plotExpression(sce, features = ids, x = "cluster", colour_by = "cluster", scales = "free")

  cat("##### Cluster", n, "\n")

  print(plt)

  cat("\n\n")

}
```

### Specific

Find specific candidate marker genes between clusters:

```{r}
res <- findMarkers(sce, sce$cluster, test.type = "binom", pval.type = "all", direction = "up")

sig <- lapply(res, subset, FDR < 0.05)

sig <- Filter(nrow, sig)
```

#### Table {.tabset}

Create a table of all significant marker genes for each cluster:

```{r, results = "asis", echo = FALSE}
for (n in names(sig)) {

  dat <- as.data.frame(sig[[n]])

  cat("##### Cluster", n, "\n")

  cat(knitr::knit_print(DT::datatable(dat)))

  cat("\n\n")

}
```

#### Heatmap {.tabset}

Plot heatmap of gene expression values for the Top 10 marker genes:

```{r, fig.height = 8, fig.width = 10, results = "asis"}
for (n in names(sig)) {
  
  dat <- sig[[n]]
  
  top <- head(dat, n = 10)
  
  ids <- rownames(top)

  cat("##### Cluster", n, "\n")
  
  plotHeatmap(sce, colour_columns_by = "cluster", features = ids, center = TRUE, symmetric = TRUE, cluster_rows = ifelse(length(ids) >= 2, TRUE, FALSE), show_colnames = FALSE)
  
  cat("\n\n")
  
}
```

#### Dimension {.tabset}

Plot reduced dimension coloured by gene expression values for the Top 10 marker genes:

```{r, fig.height = 8, fig.width = 10, results = "asis"}
for (n in names(sig)) {

  dat <- sig[[n]]

  top <- head(dat, n = 10)

  ids <- rownames(top)

  plt <- lapply(ids, function(x) plotUMAP(sce, colour_by = x) + ggtitle(x))

  cat("##### Cluster", n, "\n")

  print(patchwork::wrap_plots(plt))

  cat("\n\n")

}
```

#### Expression {.tabset}

Plot distribution of gene expression values for the Top 10 marker genes:

```{r, fig.height = 8, fig.width = 10, results = "asis"}
for (n in names(sig)) {

  dat <- sig[[n]]

  top <- head(dat, n = 10)

  ids <- rownames(top)

  plt <- plotExpression(sce, features = ids, x = "cluster", colour_by = "cluster", scales = "free")

  cat("##### Cluster", n, "\n")

  print(plt)

  cat("\n\n")

}
```

## Summary

### Output

Write experiment data:

```{r}
saveRDS(sce, file = "data/08-marker-detection.rds")
```
